+++ src/Core/Webhook/Handler/CheckoutOrderCompletedHandler.php
@@ -8,16 +8,11 @@
 namespace OxidSolutionCatalysts\PayPal\Core\Webhook\Handler;
 
 use OxidEsales\Eshop\Application\Model\Order as EshopModelOrder;
-use OxidEsales\EshopCommunity\Core\Registry;
-use OxidSolutionCatalysts\PayPal\Core\ServiceFactory;
 use OxidSolutionCatalysts\PayPalApi\Exception\ApiException;
 use OxidSolutionCatalysts\PayPal\Core\Webhook\Event;
-use OxidSolutionCatalysts\PayPal\Traits\WebhookHandlerTrait;
 
-class CheckoutOrderCompletedHandler implements HandlerInterface
+class CheckoutOrderCompletedHandler extends WebhookHandlerBase
 {
-    use WebhookHandlerTrait;
-
     /**
      * @inheritDoc
      * @throws ApiException
@@ -33,6 +28,6 @@ class CheckoutOrderCompletedHandler implements HandlerInterface
         $this->setStatus($order, $data['status'], $payPalOrderId);
         $order->markOrderPaid();
 
-        $this->cleanUpNotFinishedOrders();
+        parent::handle($event);
     }
 }